mysite/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.2
cssutils==1.0.2
Django==3.1.1
django-inlinecss==0.3.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
future==0.18.2
PyJWT==1.7.1
pynliner==0.8.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
